Work around rounding errors
When laying out a proof tree in a block with exactly the necessary space, it was possible due to floating point approximations that we would think there is not enough space, and thus lay out the premises on multiple lines. The previous fix (#16) did not work in all cases. With this new approach, we test the available space against exactly what we return when there is enough space, so there should no longer be any issue.